app/requirements.txt
fastapi==0.65.2
pydantic==1.8.2
starlette==0.14.2
python-jose~=3.3.0
SQLAlchemy==2.0.16
mysqlclient==2.1.1
python-jose==3.3.0
python-multipart==0.0.5
h11==0.13.0
cryptography==37.0.4
jwt==1.3.1
pip==21.1.2
wheel==0.36.2
rsa==4.9
pyasn1==0.4.8
ecdsa==0.18.0
pytz==2022.1
anyio==3.6.1
sniffio==1.2.0
black==22.6.0
platformdirs==2.5.2
tomli==2.0.1
pathspec==0.9.0
click==8.1.3
six==1.16.0
numpy==1.24.3
setuptools==57.0.0
cffi==1.15.1
pandas==1.4.3
openpyxl==3.1.2
python-dateutil==2.8.2
certifi==2023.5.7
requests==2.31.0
passlib==1.7.4
uvicorn==0.18.2
idna==3.3
urllib3==2.0.2
pycparser==2.21
aiofiles==23.1.0
bcrypt==4.0.0
loguru==0.6.0